# backend/docs/API_CONTRACT.yaml
openapi: 3.1.0
info:
  title: DataNova API - Plateforme Veille Réglementaire Hutchinson
  description: |
    API REST pour la veille réglementaire automatisée (CBAM, EUDR, CSRD)
    
    **Architecture:**
    - Agent 1A: Collecte documents EUR-Lex + CBAM Guidance
    - Agent 1B: Analyse de pertinence (LLM)
    - Agent 2: Évaluation d'impact détaillée
    - UI: Validation juriste + Dashboard décideur
    
  version: 1.0.0
  contact:
    name: Backend Team
    email: backend@hutchinson.com
servers:
  - url: http://localhost:8000/api
    description: Développement local
  - url: https://datanova.hutchinson.com/api
    description: Production
...
paths:
...
  /analyses:
    get:
      summary: Liste des analyses de pertinence
      description: Récupère les analyses réalisées par Agent 1B
      tags: [Analyses]
      security:
        - bearerAuth: []
      parameters:
        - name: validation_status
          in: query
          description: Statut de validation
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: is_relevant
          in: query
          description: Filtrer par pertinence
          schema:
            type: boolean
        - name: min_confidence
          in: query
          description: Confiance minimale (0-1)
          schema:
            type: number
            minimum: 0
            maximum: 1
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Liste des analyses
          content:
            application/json:
              schema:
                type: object
                properties:
                  analyses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Analysis'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
...
  /stats/analyses:
    get:
      summary: Statistiques des analyses
      description: Métriques de pertinence et validation
      tags: [Stats]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Stats analyses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisStats'

# ==================== COMPOSANTS ====================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Analysis:
      type: object
      required:
        - id
        - document_id
        - is_relevant
        - confidence
        - validation_status
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        is_relevant:
          type: boolean
          description: Document pertinent pour Hutchinson
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Niveau de confiance LLM
        matched_keywords:
          type: array
          items:
            type: string
          description: Mots-clés détectés
        matched_nc_codes:
          type: array
          items:
            type: string
          description: Codes NC correspondants
        llm_reasoning:
          type: string
          description: Explication du LLM
        validation_status:
          type: string
          enum: [pending, approved, rejected]
          description: Statut de validation juriste
        validation_comment:
          type: string
          description: Commentaire du juriste
        validated_by:
          type: string
          description: Email du validateur
        validated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    # -------------------- Impact Assessment --------------------
...
    AnalysisStats:
      type: object
      properties:
        total:
          type: integer
        relevant_count:
          type: integer
        average_confidence:
          type: number
        by_validation_status:
          type: object
          properties:
            pending:
              type: integer
            approved:
              type: integer
            rejected:
              type: integer